+++ app/src/main/java/com/erdemserhat/harmonyhaven/presentation/post_authentication/quote_main/static_card/QuoteCard.kt
@@ -1,15 +1,12 @@
-package com.erdemserhat.harmonyhaven.presentation.post_authentication.quote_main
+package com.erdemserhat.harmonyhaven.presentation.post_authentication.quote_main.static_card
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.erdemserhat.harmonyhaven.dto.responses.Quote
-import com.erdemserhat.harmonyhaven.presentation.post_authentication.quotes.FullScreenImage
 
 @Composable
-fun ImageQuote(
+fun QuoteCard(
     modifier: Modifier = Modifier,
     quoteSentence: String,
     quoteURL: String,
@@ -17,27 +14,22 @@ fun ImageQuote(
     shouldAnimate: Boolean = true,
     isFirmLogoVisible: Boolean = false
 ) {
-
     Box(modifier = modifier) {
-
-        FullScreenImage(
-            quoteURL,
+        // Display the full-screen image
+        QuoteTextCardBackground(
+            imageUrl = quoteURL,
             modifier = Modifier.align(Alignment.Center),
             shouldAnimate = shouldAnimate
         )
 
-        if (quoteSentence != "") {
-            QuoteSentence(
+        // Show the quote sentence and writer if the sentence is not empty
+        if (quoteSentence.isNotBlank()) {
+            QuoteCard(
                 quote = quoteSentence,
                 quoteWriter = quoteWriter,
                 modifier = Modifier.align(Alignment.Center),
                 isFirmLogoVisible = isFirmLogoVisible
             )
-
         }
-
-
     }
-
-
 }
