--- app/src/main/java/com/erdemserhat/harmonyhaven/presentation/post_authentication/quotes/QuotesViewModel.kt
@@ -1,39 +1,0 @@
-package com.erdemserhat.harmonyhaven.presentation.post_authentication.quotes
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.erdemserhat.harmonyhaven.domain.usecase.quote.GetQuotes
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-
-@HiltViewModel
-class QuotesViewModel @Inject constructor(
-    private val getQuotes: GetQuotes,
-) : ViewModel() {
-    private val _quotes =
-        MutableStateFlow<List<com.erdemserhat.harmonyhaven.dto.responses.Quote>>(emptyList())
-    val quotes: StateFlow<List<com.erdemserhat.harmonyhaven.dto.responses.Quote>> = _quotes
-
-    init {
-        loadQuotes()
-    }
-
-    //load notification via offset
-    private fun loadQuotes() {
-        viewModelScope.launch {
-            try {
-                _quotes.value =
-                    getQuotes.executeRequest()
-
-            } catch (e: Exception) {
-                e.printStackTrace()
-            } finally {
-
-            }
-        }
-    }
-}
